+++ drivers/ai-provider/google/factory.go
@@ -3,6 +3,8 @@ package google
 import (
 	"sync"
 
+	"github.com/eolinker/eosc/common/bean"
+
 	"github.com/eolinker/apinto/convert"
 	"github.com/eolinker/apinto/drivers"
 	"github.com/eolinker/eosc"
@@ -27,6 +29,10 @@ func NewFactory() eosc.IExtenderDriverFactory {
 
 // Create 创建驱动实例
 func Create(id, name string, v *Config, workers map[eosc.RequireId]eosc.IWorker) (eosc.IWorker, error) {
+	once.Do(func() {
+		bean.Autowired(&converterManager)
+		converterManager.Set(name, &convertFactory{})
+	})
 	_, err := checkConfig(v)
 	if err != nil {
 		return nil, err
